Allow filtering of dataset by years
This commit allows the user to filter the dataset by values available in the years column in the user interface. This also addresses #1 by removing the isMunicipal property of the model since it was merely checking to see if the table had the word “Municipal” in it to then sort by municipality.

Resolves mapc/datacommon#65