Fix Parcel Display Bug
When attempting to view the “Massachusetts Land Parcel Database (Metro Boston Region)” the rows would not display in the user interface despite the fact that the data was loading. This resolves the issue by properly returning the rows instead of the list of available years for the selected_rows computed property when the years_available is an Ember error.