Update metadata endpoint
This commit refactors the metadata endpoint call to refer to the database name instead of the schema name when summoning metadata.

This will break databrowser until MAPC/datacommon#155 is merged.

It also changes development to point to localhost for the metadata endpoints in order to allow for easier local testing.